fixed project files for 861 and 261 versions and production file
