+++ DancebotsFirmware/20200528_ATTiny261A_Flash_Memory.hex
@@ -1,0 +1,98 @@
+:1000000012C01FC01EC01DC01CC01BC01AC019C01A
+:1000100018C017C016C015C014C013C012C011C03C
+:100020008CC00FC00EC011241FBECFEDCDBF20E08D
+:10003000A0E6B0E001C01D92A736B207E1F74FD1AC
+:10004000E4C2DECF1BBA8FEE8ABB18BA8FE787BB3C
+:10005000089585B3806E85BBAC9883B7816083BFFC
+:100060008FB581608FBD89B7816089BF1DBC80B7A6
+:10007000826880BF1CBC80B7816280BFC098C29A72
+:10008000D898D99A0895682F70E080E090E0C6D1A2
+:100090002DEC3CEC4CE45FE34DD220E030E04FE24D
+:1000A00053E42AD18DD1862F0895CF93C62F87FF91
+:1000B00003C0D998D89A02C0D898D99A8F7711F0EE
+:1000C000E2DF01C080E08CBDC7FF03C0C298C09AC8
+:1000D00002C0C098C29A8C2F8F7711F0D4DF01C074
+:1000E00080E08DBDCF91089580FF02C0DA9A01C0F3
+:1000F000DA9881FF02C0DB9A01C0DB9882FF02C060
+:10010000DD9A01C0DD9883FF02C0DE9A01C0DE984F
+:1001100084FF02C0C49A01C0C49885FF02C0C69A79
+:1001200001C0C69886FF02C0DF9A01C0DF9887FF32
+:1001300002C0C59A0895C59808951F920F920FB6F0
+:100140000F9211242F933F934F935F936F937F935D
+:100150008F939F93AF93BF93EF93FF93AC9B02C09A
+:10016000AC9801C0AC9A83B322B390E0922B20915B
+:10017000650030916600BC01621B730B90936600B2
+:10018000809365006E3A26E1720758F01092640081
+:100190001092630010926200109261001092600051
+:1001A0008DC06D3A36E1730709F488C08091640010
+:1001B000883108F083C090916000913041F120F0C7
+:1001C000923009F44DC07AC09091630068327D404E
+:1001D00058F041E050E0082E01C0440F0A94EAF7BD
+:1001E000492B409363000BC021E030E0082E01C092
+:1001F000220F0A94EAF720952923209363008F5F4A
+:1002000080936400883008F459C081E026C0482FEC
+:1002100050E0909162004850510968327D4068F08A
+:1002200061E070E09B0102C0220F331F4A95E2F7A4
+:10023000A901492B409362000AC021E030E001C0CF
+:10024000220F4A95EAF720952923209362008F5FB9
+:1002500080936400803190F182E0809360002EC032
+:10026000482F50E0909161004051510968327D4023
+:1002700068F061E070E09B0102C0220F331F4A95D5
+:10028000E2F7A901492B409361000AC021E030E068
+:1002900001C0220F4A95EAF7209529232093610097
+:1002A0008F5F80936400883149F4809161001CDF86
+:1002B0006091620080916300F8DEA895FF91EF9154
+:1002C000BF91AF919F918F917F916F915F914F916E
+:1002D0003F912F910F900FBE0F901F901895B2DE97
+:1002E000B8DE9DE088E10FB6F894A89581BD0FBEF9
+:1002F00091BD7894FFCF5058BB27AA270ED0E0C0FD
+:10030000D1D030F0D6D020F031F49F3F11F41EF45C
+:10031000C6C00EF4E095E7FBBCC0E92FE2D080F345
+:10032000BA17620773078407950718F071F49EF5F2
+:10033000FAC00EF4E0950B2EBA2FA02D0B01B901D7
+:1003400090010C01CA01A0011124FF27591B99F04B
+:10035000593F50F4503E68F11A16F040A22F232F57
+:10036000342F4427585FF3CF469537952795A795A7
+:10037000F0405395C9F77EF41F16BA0B620B730B4E
+:10038000840BBAF09150A1F0FF0FBB1F661F771FBF
+:10039000881FC2F70EC0BA0F621F731F841F48F474
+:1003A000879577956795B795F7959E3F08F0B3CFFA
+:1003B0009395880F08F09927EE0F979587950895E4
+:1003C00098D088F09F5790F0B92F9927B751A0F097
+:1003D000D1F0660F771F881F991F1AF0BA95C9F7D9
+:1003E00012C0B13081F09FD0B1E008959CC0672F5A
+:1003F000782F8827B85F39F0B93FCCF38695779589
+:100400006795B395D9F73EF4909580957095619571
+:100410007F4F8F4F9F4F0895E89409C097FB3EF49C
+:1004200090958095709561957F4F8F4F9F4F992341
+:10043000A9F0F92F96E9BB279395F69587957795BF
+:100440006795B795F111F8CFFAF4BB0F11F460FF7F
+:100450001BC06F5F7F4F8F4F9F4F16C0882311F0D7
+:1004600096E911C0772321F09EE8872F762F05C0EB
+:10047000662371F096E8862F70E060E02AF09A9586
+:10048000660F771F881FDAF7880F9695879597F97B
+:10049000089597F99F6780E870E060E008959FEF06
+:1004A00080EC089500240A9416161706180609060B
+:1004B000089500240A9412161306140605060895DA
+:1004C000092E0394000C11F4882352F0BB0F40F462
+:1004D000BF2B11F460FF04C06F5F7F4F8F4F9F4FA2
+:1004E000089557FD9058440F551F59F05F3F71F024
+:1004F0004795880F97FB991F61F09F3F79F087958B
+:100500000895121613061406551FF2CF4695F1DF13
+:1005100008C0161617061806991FF1CF86957105A3
+:10052000610508940895E894BB2766277727CB01D7
+:1005300097F908950AD0C4CFB5DF28F0BADF18F0D4
+:10054000952309F0A6CFABCFEFCFCBDFA8F399234C
+:10055000D9F35523C9F3950F50E0551FAA27EE276D
+:10056000FF27BB2700240894679520F4E20FF31FB0
+:10057000B41F0A1E220F331F441FAA1F6695A9F736
+:10058000779530F4F30FB41F0A1E121E08F463951A
+:10059000330F441FAA1F221F769599F7879520F4E1
+:1005A000B40F0A1E121E631F440FAA1F221F331FFF
+:1005B0008695A9F7862F712D602D11249F575040E5
+:1005C0008AF0E1F088234AF0EE0FFF1FBB1F661F81
+:1005D000771F881F91505040A9F79E3F510570F03A
+:1005E00058CFA2CF5F3FECF3983EDCF3869577952A
+:1005F0006795B795F795E7959F5FC1F7FE2B880F35
+:0E060000911D9695879597F90895F894FFCF70
+:00000001FF
