--- DancebotsFirmware/MP3DanceBot/.gitignore
@@ -1,2 +1,0 @@
-Debug
-Release
